+++ monitor/src/main/java/com/farmaciasperuanas/pmmli/monitor/service/TransactionLogService.java
@@ -9,7 +9,7 @@ public interface TransactionLogService {
 
     List<TransactionLogDto> listarTransactionDashboard();
 
-    TransanctionDetailDto getDetailTransaction(Integer id);
+    TransanctionDetailDto getDetailTransaction(Long id);
 
     CantMaestroDto getCantidadDatosMonth();
